# laravel/laravel/my-app/storage/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: Firebase Integration API
  description: |
    # API Firebase Integration
    
    Cette API permet l'intégration complète avec Firebase, incluant:
    - Authentification avec Laravel Sanctum
    - Synchronisation bidirectionnelle Firestore/Realtime Database
    - Gestion du stockage cloud
    - Envoi de notifications push
    - Changement automatique de source de données
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
  /api/auth/register:
    post:
      tags:
        - Authentification
      summary: S'inscrire / Créer un nouveau compte
      description: Enregistrer un nouvel utilisateur et obtenir un token Sanctum
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
                name:
                  type: string
                  example: John Doe
      responses:
        '201':
          description: Compte créé avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilisateur créé"
                  token:
                    type: string
                  user:
                    type: object
        '400':
          description: Email déjà existant ou données invalides
  /api/auth/login:
    post:
      tags:
        - Authentification
      summary: Login utilisateur
      description: Authentifier un utilisateur et obtenir un token Sanctum
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  user:
                    type: object
        '401':
          description: Invalid credentials
  /api/auth/refresh:
    post:
      tags:
        - Authentification
      summary: Renouveler le token d'authentification
      description: Générer un nouveau token Sanctum pour l'utilisateur authentifié
      operationId: refreshToken
      security:
        - sanctum: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
        '401':
          description: Unauthenticated
...
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Laravel Sanctum token
...